Validate config file
Parse and check configuation such that the file is deployed if and only
if it is valid, i.e. invalid configurations (for example due to typoes
in any of the weakly types class parameters) will cause puppet to fail
instead of shipping a file that is known to produce doas(1) errors.

See https://man.openbsd.org/doas.1#C